# openapi.yaml
openapi: 3.1.0
info:
  title: AgriCRM API
  description: Farmer, FPO & Sugar Mill CRM for Theta Analytics
  version: 1.0.0
paths:
...
  /api/v1/reminder-policies/:
    get:
      tags:
      - delivery
      summary: Reminder Policy List
      operationId: reminder_policy_list_api_v1_reminder_policies__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ReminderPolicyOut'
                type: array
                title: Response Reminder Policy List Api V1 Reminder Policies  Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - delivery
      summary: Reminder Policy Create
      description: "Create a policy. \U0001F534 Autosend is off; enabling it is a\
        \ separate act."
      operationId: reminder_policy_create_api_v1_reminder_policies__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderPolicyCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderPolicyOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/reminder-policies/{policy_id}/autosend/:
    post:
      tags:
      - delivery
      summary: Reminder Policy Autosend
      description: "\U0001F534 Let a scheduled run send without a person, or stop\
        \ it.\n\nSeparately permissioned, and it records who, when and why. Enabling\n\
        requires a ceiling above zero \u2014 the DDL refuses the combination of\n\"\
        enabled\" with no limit and no named enabler, so this cannot be half-done."
      operationId: reminder_policy_autosend_api_v1_reminder_policies__policy_id__autosend__post
      security:
      - HTTPBearer: []
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Policy Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutosendRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderPolicyOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reminder-runs/preview/:
    post:
      tags:
      - delivery
      summary: Reminder Run Preview
      description: "Build a batch preview. Sends nothing.\n\n\U0001F534 The response\
        \ carries `skipped` beside `candidates`. A preview showing\nonly who would\
        \ be contacted hides the decisions worth reviewing \u2014 the\nopt-outs, the\
        \ frequency caps, the promises to pay."
      operationId: reminder_run_preview_api_v1_reminder_runs_preview__post
      security:
      - HTTPBearer: []
      parameters:
      - name: policy
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Policy
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderRunOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reminder-runs/{run_id}/confirm/:
    post:
      tags:
      - delivery
      summary: Reminder Run Confirm
      description: "Confirm a frozen preview and queue every candidate.\n\n\U0001F534\
        \ The preview is rebuilt and re-hashed before anything is queued. A\npayment\
        \ that arrived, an opt-out that landed or an invoice that was\ncancelled since\
        \ the screen was rendered changes the hash, and the whole run\nis refused\
        \ rather than partly sent."
      operationId: reminder_run_confirm_api_v1_reminder_runs__run_id__confirm__post
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderConfirm'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderRunResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AutosendRequest:
      properties:
        enabled:
          type: boolean
          title: Enabled
        max_per_run:
          type: integer
          title: Max Per Run
          default: 0
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      type: object
      required:
      - enabled
      - reason
      title: AutosendRequest
      description: "Enable or disable unattended sending for a policy.\n\n\U0001F534\
        \ Enabling requires a limit above zero and a reason. Both are recorded,\n\
        because \"who decided the machine could message customers unattended, and\n\
        what did they think the ceiling was\" is the question afterwards."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReminderConfirm:
      properties:
        preview_sha256:
          type: string
          maxLength: 64
          minLength: 64
          title: Preview Sha256
      type: object
      required:
      - preview_sha256
      title: ReminderConfirm
    ReminderPolicyCreate:
      properties:
        billing_entity:
          type: string
          format: uuid
          title: Billing Entity
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        channel:
          type: string
          title: Channel
          default: email
        trigger_days:
          items:
            type: integer
          type: array
          title: Trigger Days
          default:
          - 7
          - 15
          - 30
        template_body:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Body
        quiet_hour_start:
          type: integer
          title: Quiet Hour Start
          default: 20
        quiet_hour_end:
          type: integer
          title: Quiet Hour End
          default: 9
        timezone:
          type: string
          title: Timezone
          default: Asia/Kolkata
        min_days_between:
          type: integer
          title: Min Days Between
          default: 7
        max_per_invoice:
          type: integer
          title: Max Per Invoice
          default: 4
      type: object
      required:
      - billing_entity
      - name
      title: ReminderPolicyCreate
    ReminderPolicyOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        billing_entity:
          type: string
          format: uuid
          title: Billing Entity
        name:
          type: string
          title: Name
        channel:
          type: string
          title: Channel
        trigger_days:
          items:
            type: integer
          type: array
          title: Trigger Days
        template_version:
          type: string
          title: Template Version
        quiet_hour_start:
          type: integer
          title: Quiet Hour Start
        quiet_hour_end:
          type: integer
          title: Quiet Hour End
        timezone:
          type: string
          title: Timezone
        min_days_between:
          type: integer
          title: Min Days Between
        max_per_invoice:
          type: integer
          title: Max Per Invoice
        autosend_enabled:
          type: boolean
          title: Autosend Enabled
        autosend_max_per_run:
          type: integer
          title: Autosend Max Per Run
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - id
      - billing_entity
      - name
      - channel
      - trigger_days
      - template_version
      - quiet_hour_start
      - quiet_hour_end
      - timezone
      - min_days_between
      - max_per_invoice
      - autosend_enabled
      - autosend_max_per_run
      - is_active
      title: ReminderPolicyOut
    ReminderRunOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        policy_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Policy Id
        status:
          type: string
          title: Status
        preview_sha256:
          type: string
          title: Preview Sha256
        candidates:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Candidates
        skipped:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Skipped
        expires_at:
          type: string
          title: Expires At
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - policy_id
      - status
      - preview_sha256
      - candidates
      - skipped
      - expires_at
      - created_at
      title: ReminderRunOut
    ReminderRunResult:
      properties:
        run:
          $ref: '#/components/schemas/ReminderRunOut'
        queued:
          type: integer
          title: Queued
        reminders:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Reminders
      type: object
      required:
      - run
      - queued
      - reminders
      title: ReminderRunResult
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer